sensor-producer/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
frozenlist==1.5.0
idna==3.10
kafka-python==2.1.5
multidict==6.3.2
numpy==2.2.4
pandas==2.2.3
propcache==0.3.1
psutil==7.0.0
psycopg==3.2.6
puremagic==1.28
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-weather==2.0.7
pytz==2025.2
six==1.17.0
typing_extensions==4.13.1
tzdata==2025.2
yarl==1.19.0
